Début du développement de l'exploration par plus proche frontière
